Handle log files that change names
When service is restarted the log file changes name - that means we want
to reset the state of the logfile so we can read all of it.

Note this won't process all logs if service restarted multiple times in
a day. This is not the 'correct' way to do things